requirements.txt
pandas
numpy
scikit-learn
streamlit
folium
streamlit-folium
pdf2image
matplotlib
supabase
python-dotenv